refactor(web): extract memory-provider mixins from web_server.py (shard s2, logger/_impl fixes)
