+++ hermes_cli/web_server.py
@@ -100,6 +100,48 @@ from gateway.status import (
 )
 from utils import env_var_enabled
 
+from hermes_cli.web_routers.memory_provider_native import (
+    _coerce_bool,
+    _coerce_schema_field,
+    _discover_memory_provider_statuses_impl as _discover_memory_provider_statuses,
+    _env_lookup,
+    _field_default,
+    _field_is_set,
+    _field_value,
+    _field_visible,
+    _memory_provider_is_configured,
+    _memory_provider_payload,
+    _normalize_memory_provider_schema,
+    _public_memory_provider_field,
+    _read_json_file,
+    _read_memory_provider_existing_values,
+    _require_memory_provider_ready,
+    _require_valid_memory_provider_name,
+    _save_memory_provider_native_config,
+    _write_memory_provider_config_values,
+)
+from hermes_cli.web_routers.memory_provider_setup import (
+    _command_result,
+    _dependency_importable_impl as _dependency_importable,
+    _install_memory_provider_external_dependencies,
+    _install_memory_provider_pip_dependencies,
+    _install_memory_provider_setup,
+    _load_memory_provider,
+    _memory_provider_dependencies_installed,
+    _memory_provider_dependency_package,
+    _memory_provider_import_name,
+    _memory_provider_label,
+    _memory_provider_manifest,
+    _memory_provider_setup_env,
+    _memory_provider_setup_info,
+    _memory_provider_setup_manifest,
+    _normalize_memory_provider_name,
+    _run_setup_command,
+    _string_list,
+    _trim_setup_output,
+)
+
+
 try:
     from fastapi import (
         FastAPI, File, Form, HTTPException, Query, Request, UploadFile,
@@ -5172,833 +5214,6 @@ def _update_memory_provider_config(provider: ProviderConfigSchema, values: Dict[
     if memory_config.get("provider") != provider.name:
         memory_config["provider"] = provider.name
         save_config(config)
-
-
-def _memory_provider_label(name: str) -> str:
-    return name.replace("_", " ").replace("-", " ").title()
-
-
-def _normalize_memory_provider_name(name: Any) -> str:
-    provider = str(name or "").strip()
-    if provider.lower() in {"built-in", "builtin", "none"}:
-        return ""
-    return provider
-
-
-def _load_memory_provider(name: str):
-    try:
-        from plugins.memory import load_memory_provider
-
-        return load_memory_provider(name)
-    except Exception:
-        _log.debug("Failed to load memory provider %s", name, exc_info=True)
-        return None
-
-
-def _memory_provider_manifest(name: str) -> Dict[str, Any]:
-    try:
-        from plugins.memory import find_provider_dir
-
-        provider_dir = find_provider_dir(name)
-        if provider_dir is None:
-            return {}
-        manifest_path = provider_dir / "plugin.yaml"
-        if not manifest_path.exists():
-            return {}
-        with manifest_path.open(encoding="utf-8-sig") as handle:
-            manifest = yaml.safe_load(handle) or {}
-        return manifest if isinstance(manifest, dict) else {}
-    except Exception:
-        _log.debug("Failed to read memory provider manifest for %s", name, exc_info=True)
-        return {}
-
-
-def _string_list(value: Any) -> List[str]:
-    if not isinstance(value, list):
-        return []
-    return [str(item).strip() for item in value if str(item).strip()]
-
-
-def _memory_provider_setup_manifest(name: str) -> Dict[str, Any]:
-    manifest = _memory_provider_manifest(name)
-    external_dependencies: List[Dict[str, str]] = []
-    for raw in manifest.get("external_dependencies") or []:
-        if not isinstance(raw, dict):
-            continue
-        dep = {
-            "name": str(raw.get("name") or "").strip(),
-            "install": str(raw.get("install") or "").strip(),
-            "check": str(raw.get("check") or "").strip(),
-        }
-        if dep["name"] or dep["install"] or dep["check"]:
-            external_dependencies.append(dep)
-
-    return {
-        "pip_dependencies": _string_list(manifest.get("pip_dependencies")),
-        "external_dependencies": external_dependencies,
-        "required_env": _string_list(manifest.get("requires_env")),
-    }
-
-
-def _memory_provider_setup_info(name: str) -> Dict[str, Any]:
-    setup = _memory_provider_setup_manifest(name)
-    setup["dependencies_installed"] = _memory_provider_dependencies_installed(setup)
-    return setup
-
-
-_MEMORY_PROVIDER_IMPORT_NAMES = {
-    "honcho-ai": "honcho",
-    "mem0ai": "mem0",
-    "hindsight-client": "hindsight_client",
-    "hindsight-all": "hindsight",
-}
-
-
-def _memory_provider_dependency_package(dep: str) -> str:
-    return re.split(r"[\[<>=!~;]", dep, maxsplit=1)[0].strip()
-
-
-def _memory_provider_import_name(dep: str) -> str:
-    package = _memory_provider_dependency_package(dep)
-    return _MEMORY_PROVIDER_IMPORT_NAMES.get(package, package.replace("-", "_"))
-
-
-def _dependency_importable(dep: str) -> bool:
-    import_name = _memory_provider_import_name(dep)
-    if not import_name:
-        return False
-    try:
-        __import__(import_name)
-        return True
-    except ImportError:
-        return False
-
-
-def _trim_setup_output(value: Optional[str], limit: int = 4000) -> str:
-    text = str(value or "")
-    if len(text) <= limit:
-        return text
-    return f"{text[:limit]}\n... truncated ..."
-
-
-def _memory_provider_setup_env() -> Dict[str, str]:
-    # External package-manager child (npm/uv/pip): exact env preservation —
-    # scrubbing or HOME rewriting could break user tool auth/config.
-    from tools.environments.local import build_subprocess_env
-    env = build_subprocess_env(scrub_secrets=False, inherit_profile_home=False)
-    home = Path.home()
-    extra_bins = [
-        home / ".brv-cli" / "bin",
-        home / ".local" / "bin",
-        home / ".npm-global" / "bin",
-        Path("/usr/local/bin"),
-    ]
-    existing_path = env.get("PATH", "")
-    prefix = os.pathsep.join(str(path) for path in extra_bins if path.exists())
-    if prefix:
-        env["PATH"] = prefix + os.pathsep + existing_path
-    return env
-
-
-def _command_result(
-    *,
-    kind: str,
-    name: str,
-    status: str,
-    command: str = "",
-    completed: Optional[subprocess.CompletedProcess] = None,
-    error: Optional[str] = None,
-) -> Dict[str, Any]:
-    return {
-        "kind": kind,
-        "name": name,
-        "status": status,
-        "command": command,
-        "returncode": None if completed is None else completed.returncode,
-        "stdout": "" if completed is None else _trim_setup_output(completed.stdout),
-        "stderr": _trim_setup_output(error or ("" if completed is None else completed.stderr)),
-    }
-
-
-def _run_setup_command(
-    command: Any,
-    *,
-    display: str,
-    shell: bool = False,
-    timeout: int = 180,
-) -> subprocess.CompletedProcess:
-    return subprocess.run(
-        command,
-        shell=shell,
-        executable="/bin/bash" if shell else None,
-        env=_memory_provider_setup_env(),
-        capture_output=True,
-        text=True,
-        # Lossy UTF-8 decode — setup tools emit UTF-8; never let a
-        # locale-mismatched byte raise in the reader thread (#52649).
-        encoding="utf-8",
-        errors="replace",
-        timeout=timeout,
-        check=False,
-    )
-
-
-def _memory_provider_dependencies_installed(setup: Dict[str, Any]) -> bool:
-    pip_dependencies = _string_list(setup.get("pip_dependencies"))
-    external_dependencies = setup.get("external_dependencies") or []
-
-    pip_ok = all(_dependency_importable(dep) for dep in pip_dependencies)
-    external_ok = True
-    for dep in external_dependencies:
-        if not isinstance(dep, dict):
-            continue
-        check_cmd = str(dep.get("check") or "").strip()
-        install_cmd = str(dep.get("install") or "").strip()
-        if not check_cmd:
-            if install_cmd:
-                external_ok = False
-            continue
-        try:
-            completed = _run_setup_command(
-                shlex.split(check_cmd),
-                display=check_cmd,
-                timeout=20,
-            )
-        except Exception:
-            external_ok = False
-            continue
-        if completed.returncode != 0:
-            external_ok = False
-
-    return pip_ok and external_ok
-
-
-def _install_memory_provider_pip_dependencies(dependencies: List[str]) -> List[Dict[str, Any]]:
-    missing = [dep for dep in dependencies if not _dependency_importable(dep)]
-    if not dependencies:
-        return []
-    if not missing:
-        return [
-            _command_result(kind="pip", name=", ".join(dependencies), status="already_installed")
-        ]
-
-    # Route through the lazy-install pipeline (tools.lazy_deps.install_specs)
-    # instead of shelling out to pip against sys.executable directly. That
-    # pipeline is environment-aware: on hosted/immutable images the agent venv
-    # under /opt/hermes is sealed read-only, and installs must be redirected
-    # to the writable durable target on the data volume
-    # (HERMES_LAZY_INSTALL_TARGET, e.g. /opt/data/lazy-packages) — the same
-    # path every lazy backend already uses. A direct `pip install --python
-    # sys.executable` on those images fails with a permission error (NS-605).
-    # install_specs also activates the target on sys.path post-install so the
-    # availability recheck below sees the new packages without a restart.
-    try:
-        from tools.lazy_deps import install_specs
-
-        outcome = install_specs(missing, timeout=240)
-    except Exception as exc:
-        return [
-            _command_result(
-                kind="pip",
-                name=", ".join(missing),
-                status="failed",
-                error=str(exc),
-            )
-        ]
-
-    if outcome.blocked:
-        return [
-            _command_result(
-                kind="pip",
-                name=", ".join(missing),
-                status="failed",
-                command=outcome.command,
-                error=outcome.reason,
-            )
-        ]
-
-    return [
-        _command_result(
-            kind="pip",
-            name=", ".join(missing),
-            status="installed" if outcome.ok else "failed",
-            command=outcome.command,
-            completed=subprocess.CompletedProcess(
-                args=outcome.command,
-                returncode=0 if outcome.ok else 1,
-                stdout=outcome.stdout,
-                stderr=outcome.stderr,
-            ),
-        )
-    ]
-
-
-def _install_memory_provider_external_dependencies(
-    dependencies: List[Dict[str, str]],
-) -> List[Dict[str, Any]]:
-    results: List[Dict[str, Any]] = []
-    for dep in dependencies:
-        name = dep.get("name") or "dependency"
-        check_cmd = dep.get("check") or ""
-        install_cmd = dep.get("install") or ""
-
-        if check_cmd:
-            try:
-                check = _run_setup_command(
-                    shlex.split(check_cmd),
-                    display=check_cmd,
-                    timeout=20,
-                )
-            except Exception as exc:
-                results.append(
-                    _command_result(
-                        kind="external_check",
-                        name=name,
-                        status="missing" if install_cmd else "failed",
-                        command=check_cmd,
-                        error=str(exc),
-                    )
-                )
-            else:
-                if check.returncode == 0:
-                    results.append(
-                        _command_result(
-                            kind="external_check",
-                            name=name,
-                            status="already_installed",
-                            command=check_cmd,
-                            completed=check,
-                        )
-                    )
-                    continue
-                results.append(
-                    _command_result(
-                        kind="external_check",
-                        name=name,
-                        status="missing" if install_cmd else "failed",
-                        command=check_cmd,
-                        completed=check,
-                    )
-                )
-
-            if not install_cmd:
-                continue
-
-        if install_cmd:
-            try:
-                install = _run_setup_command(
-                    install_cmd,
-                    display=install_cmd,
-                    shell=True,
-                    timeout=300,
-                )
-            except Exception as exc:
-                results.append(
-                    _command_result(
-                        kind="external_install",
-                        name=name,
-                        status="failed",
-                        command=install_cmd,
-                        error=str(exc),
-                    )
-                )
-                continue
-
-            results.append(
-                _command_result(
-                    kind="external_install",
-                    name=name,
-                    status="installed" if install.returncode == 0 else "failed",
-                    command=install_cmd,
-                    completed=install,
-                )
-            )
-
-            if check_cmd and install.returncode == 0:
-                try:
-                    post_check = _run_setup_command(
-                        shlex.split(check_cmd),
-                        display=check_cmd,
-                        timeout=20,
-                    )
-                    results.append(
-                        _command_result(
-                            kind="external_check",
-                            name=name,
-                            status="verified" if post_check.returncode == 0 else "failed",
-                            command=check_cmd,
-                            completed=post_check,
-                        )
-                    )
-                except Exception as exc:
-                    results.append(
-                        _command_result(
-                            kind="external_check",
-                            name=name,
-                            status="failed",
-                            command=check_cmd,
-                            error=str(exc),
-                        )
-                    )
-
-    return results
-
-
-def _install_memory_provider_setup(name: str) -> Dict[str, Any]:
-    provider = _load_memory_provider(name)
-    manifest = _memory_provider_manifest(name)
-    if provider is None and not manifest:
-        raise HTTPException(status_code=404, detail=f"Unknown memory provider: {name}")
-
-    setup = _memory_provider_setup_manifest(name)
-    results = []
-    results.extend(_install_memory_provider_pip_dependencies(setup["pip_dependencies"]))
-    results.extend(
-        _install_memory_provider_external_dependencies(setup["external_dependencies"])
-    )
-
-    if not results:
-        results.append(
-            _command_result(
-                kind="setup",
-                name=name,
-                status="no_declared_steps",
-            )
-        )
-
-    ok = all(result["status"] not in {"failed"} for result in results)
-    statuses = {row["name"]: row for row in _discover_memory_provider_statuses()}
-    return {
-        "ok": ok,
-        "provider": name,
-        "results": results,
-        "status": statuses.get(name),
-    }
-
-
-def _normalize_memory_provider_schema(name: str, provider: Any) -> List[Dict[str, Any]]:
-    raw_schema: List[Dict[str, Any]] = []
-    if provider is not None and hasattr(provider, "get_config_schema"):
-        try:
-            raw = provider.get_config_schema()
-            if isinstance(raw, list):
-                raw_schema = [field for field in raw if isinstance(field, dict)]
-        except Exception:
-            _log.warning("Failed to read memory provider schema for %s", name, exc_info=True)
-
-    fields: List[Dict[str, Any]] = []
-    for raw in raw_schema:
-        key = str(raw.get("key") or "").strip()
-        if not key:
-            continue
-
-        choices = raw.get("choices") or raw.get("options") or []
-        if not isinstance(choices, list):
-            choices = []
-
-        explicit_kind = str(raw.get("kind") or raw.get("type") or "").strip().lower()
-        if raw.get("secret"):
-            kind = "secret"
-        elif choices:
-            kind = "select"
-        elif explicit_kind in {"bool", "boolean"} or isinstance(raw.get("default"), bool):
-            kind = "boolean"
-        elif explicit_kind in {"int", "integer"} or (
-            isinstance(raw.get("default"), int) and not isinstance(raw.get("default"), bool)
-        ):
-            kind = "integer"
-        elif explicit_kind in {"float", "number"} or isinstance(raw.get("default"), float):
-            kind = "number"
-        else:
-            kind = "text"
-
-        options = []
-        for choice in choices:
-            value = str(choice)
-            options.append({"value": value, "label": value, "description": ""})
-
-        description = str(raw.get("description") or "")
-        fields.append({
-            "key": key,
-            "label": str(raw.get("label") or key.replace("_", " ").title()),
-            "kind": kind,
-            "description": description,
-            "placeholder": str(raw.get("placeholder") or ""),
-            "required": bool(raw.get("required", False)),
-            "default": raw.get("default", ""),
-            "options": options,
-            "url": str(raw.get("url") or ""),
-            "when": raw.get("when") if isinstance(raw.get("when"), dict) else None,
-            "minimum": raw.get("minimum"),
-            "maximum": raw.get("maximum"),
-            "step": raw.get("step"),
-            "_env_key": str(raw.get("env_var") or "") or None,
-        })
-
-    return fields
-
-
-def _read_json_file(path: Path) -> Dict[str, Any]:
-    if not path.exists():
-        return {}
-    try:
-        data = json.loads(path.read_text(encoding="utf-8"))
-    except Exception:
-        _log.debug("Failed to read JSON config from %s", path, exc_info=True)
-        return {}
-    return data if isinstance(data, dict) else {}
-
-
-def _read_memory_provider_existing_values(name: str) -> Dict[str, Any]:
-    """Best-effort read of existing provider config across legacy/native stores."""
-
-    hermes_home = get_hermes_home()
-    values: Dict[str, Any] = {}
-
-    # Common native provider stores.
-    for path in (
-        hermes_home / f"{name}.json",
-        hermes_home / name / "config.json",
-    ):
-        values.update(_read_json_file(path))
-
-    try:
-        cfg = load_config()
-    except Exception:
-        cfg = {}
-
-    memory_cfg = cfg.get("memory") if isinstance(cfg, dict) else {}
-    if isinstance(memory_cfg, dict):
-        provider_cfg = memory_cfg.get(name)
-        if isinstance(provider_cfg, dict):
-            values.update(provider_cfg)
-        legacy_cfg = memory_cfg.get("provider_config")
-        if isinstance(legacy_cfg, dict):
-            values = {**legacy_cfg, **values}
-
-    # Holographic stores under plugins.hermes-memory-store.
-    plugins_cfg = cfg.get("plugins") if isinstance(cfg, dict) else {}
-    if name == "holographic" and isinstance(plugins_cfg, dict):
-        holographic_cfg = plugins_cfg.get("hermes-memory-store")
-        if isinstance(holographic_cfg, dict):
-            values.update(holographic_cfg)
-
-    return values
-
-
-def _env_lookup(env_key: Optional[str]) -> str:
-    if not env_key:
-        return ""
-    env_on_disk = load_env()
-    return str(env_on_disk.get(env_key) or os.environ.get(env_key) or "")
-
-
-def _coerce_bool(value: Any, *, default: bool = False) -> bool:
-    if isinstance(value, bool):
-        return value
-    if value is None or value == "":
-        return default
-    if isinstance(value, (int, float)):
-        return bool(value)
-    text = str(value).strip().lower()
-    if text in {"1", "true", "yes", "on"}:
-        return True
-    if text in {"0", "false", "no", "off"}:
-        return False
-    raise ValueError(f"Invalid boolean value: {value}")
-
-
-def _field_default(field: Dict[str, Any]) -> Any:
-    default = field.get("default", "")
-    if field["kind"] == "boolean":
-        return _coerce_bool(default, default=False)
-    return default
-
-
-def _field_value(field: Dict[str, Any], data: Dict[str, Any]) -> Any:
-    if field["kind"] == "secret":
-        return ""
-
-    value = data.get(field["key"])
-    if value in (None, ""):
-        value = _env_lookup(field.get("_env_key"))
-    if value in (None, ""):
-        value = _field_default(field)
-
-    if field["kind"] == "select":
-        allowed = {opt["value"] for opt in field.get("options", [])}
-        value = str(value)
-        return value if value in allowed else str(_field_default(field))
-    if field["kind"] == "boolean":
-        return _coerce_bool(value, default=_coerce_bool(_field_default(field), default=False))
-    return str(value)
-
-
-def _field_is_set(field: Dict[str, Any], data: Dict[str, Any]) -> bool:
-    if field["kind"] == "secret":
-        return bool(_env_lookup(field.get("_env_key")) or data.get(field["key"]))
-    value = _field_value(field, data)
-    return value not in (None, "")
-
-
-def _field_visible(
-    field: Dict[str, Any],
-    data: Dict[str, Any],
-    fields_by_key: Optional[Dict[str, Dict[str, Any]]] = None,
-) -> bool:
-    when = field.get("when")
-    if not isinstance(when, dict) or not when:
-        return True
-    for dep_key, expected in when.items():
-        dep_field = (fields_by_key or {}).get(str(dep_key)) or {
-            "key": str(dep_key),
-            "kind": "text",
-            "default": "",
-            "_env_key": None,
-        }
-        actual = _field_value(dep_field, data)
-        if str(actual) != str(expected):
-            return False
-    return True
-
-
-def _public_memory_provider_field(field: Dict[str, Any], data: Dict[str, Any]) -> Dict[str, Any]:
-    entry = {
-        "key": field["key"],
-        "label": field["label"],
-        "kind": field["kind"],
-        "description": field["description"],
-        "placeholder": field["placeholder"],
-        "required": field["required"],
-        "value": "" if field["kind"] == "secret" else _field_value(field, data),
-        "is_set": _field_is_set(field, data),
-        "options": field.get("options", []),
-        "url": field.get("url", ""),
-        "when": field.get("when"),
-        "minimum": field.get("minimum"),
-        "maximum": field.get("maximum"),
-        "step": field.get("step"),
-    }
-    return entry
-
-
-def _memory_provider_payload(name: str, provider: Any) -> Dict[str, Any]:
-    data = _read_memory_provider_existing_values(name)
-    fields = [
-        _public_memory_provider_field(field, data)
-        for field in _normalize_memory_provider_schema(name, provider)
-    ]
-    return {
-        "name": name,
-        "label": _memory_provider_label(name),
-        "fields": fields,
-        "setup": _memory_provider_setup_info(name),
-    }
-
-
-def _coerce_schema_field(field: Dict[str, Any], raw: Any) -> Any:
-    if field["kind"] == "boolean":
-        return _coerce_bool(raw, default=_coerce_bool(_field_default(field), default=False))
-
-    if field["kind"] in {"integer", "number"}:
-        value = raw if raw is not None and raw != "" else _field_default(field)
-        try:
-            if isinstance(value, bool):
-                raise ValueError
-            parsed = float(value)
-            if not math.isfinite(parsed):
-                raise ValueError
-            if field["kind"] == "integer":
-                if not parsed.is_integer():
-                    raise ValueError
-                result: int | float = int(parsed)
-            else:
-                result = parsed
-        except (TypeError, ValueError, OverflowError) as exc:
-            raise ValueError(f"Invalid numeric value for '{field['key']}'") from exc
-
-        minimum = field.get("minimum")
-        maximum = field.get("maximum")
-        if minimum is not None and result < minimum:
-            raise ValueError(f"'{field['key']}' must be at least {minimum}")
-        if maximum is not None and result > maximum:
-            raise ValueError(f"'{field['key']}' must be at most {maximum}")
-        return result
-
-    value = str(raw if raw is not None else "").strip()
-    if field["kind"] == "select":
-        if not value:
-            value = str(_field_default(field))
-        allowed = {opt["value"] for opt in field.get("options", [])}
-        if value not in allowed:
-            raise ValueError(f"Invalid value for '{field['key']}'")
-        return value
-
-    return value or _field_default(field)
-
-
-def _save_memory_provider_native_config(name: str, provider: Any, values: Dict[str, Any]) -> None:
-    if provider is not None and hasattr(provider, "save_config"):
-        try:
-            from agent.memory_provider import MemoryProvider as _BaseMemoryProvider
-        except Exception:
-            provider.save_config(values, str(get_hermes_home()))
-            return
-        if type(provider).save_config is not _BaseMemoryProvider.save_config:
-            provider.save_config(values, str(get_hermes_home()))
-            return
-
-    cfg = load_config()
-    memory_cfg = cfg.get("memory")
-    if not isinstance(memory_cfg, dict):
-        memory_cfg = {}
-        cfg["memory"] = memory_cfg
-    current = memory_cfg.get(name)
-    if not isinstance(current, dict):
-        current = {}
-    current.update(values)
-    memory_cfg[name] = current
-    save_config(cfg)
-
-
-def _memory_provider_is_configured(name: str, provider: Any) -> bool:
-    data = _read_memory_provider_existing_values(name)
-    fields = _normalize_memory_provider_schema(name, provider)
-    fields_by_key = {field["key"]: field for field in fields}
-    visible_fields = [
-        field for field in fields if _field_visible(field, data, fields_by_key)
-    ]
-    required_fields = [field for field in visible_fields if field.get("required")]
-    if not required_fields:
-        return True
-    return all(_field_is_set(field, data) for field in required_fields)
-
-
-def _discover_memory_provider_statuses() -> List[Dict[str, Any]]:
-    discovered: Dict[str, Dict[str, Any]] = {}
-    try:
-        from plugins.memory import discover_memory_providers
-
-        for name, description, available in discover_memory_providers():
-            discovered[str(name)] = {
-                "name": str(name),
-                "description": str(description or ""),
-                "available": bool(available),
-                "missing": False,
-            }
-    except Exception:
-        _log.exception("discover_memory_providers failed")
-
-    cfg = load_config()
-    active = ""
-    mem = cfg.get("memory")
-    if isinstance(mem, dict):
-        active = _normalize_memory_provider_name(mem.get("provider"))
-    if active and active not in discovered:
-        discovered[active] = {
-            "name": active,
-            "description": "Configured provider was not found.",
-            "available": False,
-            "missing": True,
-        }
-
-    providers: List[Dict[str, Any]] = []
-    for name in sorted(discovered):
-        row = discovered[name]
-        provider = None if row["missing"] else _load_memory_provider(name)
-        setup = _memory_provider_setup_info(name)
-        configured = False if row["missing"] else _memory_provider_is_configured(name, provider)
-        schema_fields = [] if row["missing"] else _normalize_memory_provider_schema(name, provider)
-        if row["missing"]:
-            status = "missing"
-        elif not row["available"] and not setup.get("dependencies_installed", True):
-            status = "unavailable"
-        elif not configured:
-            status = "needs_config"
-        elif not row["available"] and schema_fields:
-            status = "needs_config"
-        elif not row["available"]:
-            status = "unavailable"
-        else:
-            status = "ready"
-        providers.append({
-            "name": name,
-            "description": row["description"],
-            "available": row["available"],
-            "configured": configured,
-            "status": status,
-            "setup": setup,
-        })
-    return providers
-
-
-def _require_memory_provider_ready(name: str) -> None:
-    if not name:
-        return
-    statuses = {row["name"]: row for row in _discover_memory_provider_statuses()}
-    row = statuses.get(name)
-    if row is None:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Unknown memory provider '{name}'.",
-        )
-    if row["status"] != "ready":
-        raise HTTPException(
-            status_code=400,
-            detail=(
-                f"Memory provider '{name}' is not ready "
-                f"({row['status'].replace('_', ' ')}). Configure it in the dashboard first."
-            ),
-        )
-
-
-def _write_memory_provider_config_values(
-    name: str,
-    provider: Any,
-    values: Dict[str, Any],
-) -> None:
-    existing = _read_memory_provider_existing_values(name)
-    fields = _normalize_memory_provider_schema(name, provider)
-    fields_by_key = {field["key"]: field for field in fields}
-    config_values: Dict[str, Any] = {}
-    secrets: Dict[str, str] = {}
-
-    for field in fields:
-        if not _field_visible(field, {**existing, **config_values}, fields_by_key):
-            continue
-
-        if field["kind"] == "secret":
-            submitted = str(values.get(field["key"]) or "").strip()
-            if submitted and field.get("_env_key"):
-                secrets[str(field["_env_key"])] = submitted
-            continue
-
-        raw = (
-            values[field["key"]]
-            if field["key"] in values
-            else existing.get(field["key"], _field_default(field))
-        )
-        config_values[field["key"]] = _coerce_schema_field(field, raw)
-
-    _save_memory_provider_native_config(name, provider, config_values)
-
-    for env_key, secret in secrets.items():
-        save_env_value(env_key, secret)
-
-
-_MEMORY_PROVIDER_NAME_RE = re.compile(r"^[A-Za-z0-9][A-Za-z0-9_-]{0,63}$")
-
-
-def _require_valid_memory_provider_name(name: str) -> None:
-    """Reject provider names that could traverse outside the plugin dirs.
-
-    ``name`` is interpolated into filesystem paths by ``find_provider_dir()``
-    and gates which plugin manifest's setup commands run. A strict charset
-    allowlist (no path separators, no dots) makes traversal impossible
-    regardless of how the downstream lookup evolves.
-    """
-    if not _MEMORY_PROVIDER_NAME_RE.fullmatch(name or ""):
-        raise HTTPException(status_code=404, detail=f"Unknown memory provider: {name}")
 
 
 @app.get("/api/memory/providers/{name}/config")
